% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\docType{data}
\name{a4_height_mm}
\alias{a4_height_mm}
\title{A4 useable height.}
\format{An object of class \code{numeric} of length 1.}
\usage{
a4_height_mm
}
\value{
A numeric value.
}
\description{
The height of useable space within an a4 sheet.
}
\keyword{datasets}
