% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{add_tip}
\alias{add_tip}
\title{Add a quick tooltip text column to data.}
\usage{
add_tip(data, vars_vctr)
}
\arguments{
\item{data}{A tibble or dataframe. Required input.}

\item{vars_vctr}{A vector of quoted variables to include in the tooltip. Required input.}
}
\value{
A vector of labels.
}
\description{
Add a column of tooltip text which is automatically created based on column names and values.
}
