% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{colnames_to_snake_to_sentence}
\alias{colnames_to_snake_to_sentence}
\title{Convert column names to snake case and then to sentence case.}
\usage{
colnames_to_snake_to_sentence(data)
}
\arguments{
\item{data}{The number of digits to round the legend labels.}
}
\value{
A numeric value.
}
\description{
Convert colnames to snakecase and then to sentence case to present colnames visually.
}
\keyword{internal}
