% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_borders}
\alias{example_borders}
\title{Example sf object of the New Zealand coastline.}
\format{
An \code{sf} object.
}
\usage{
example_borders
}
\description{
Example sf object of the New Zealand coastline used to demonstrate adding borders to maps.
}
\examples{
gg_sf_col(example_point, col_var = trend_category, borders = example_borders)
}
\keyword{datasets}
