% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_point}
\alias{example_point}
\title{Example sf point object.}
\format{
An \code{sf} object.
}
\usage{
example_point
}
\description{
Example sf point object.
}
\examples{
gg_sf_col(example_polygon, col_var = density, borders = example_borders)
}
\keyword{datasets}
