% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_sf_nz_livestock}
\alias{example_sf_nz_livestock}
\title{Example sf object of New Zealand livestock.}
\format{An \code{sf} object.}
\usage{
example_sf_nz_livestock
}
\description{
Example sf object of New Zealand livestock.
}
\examples{
example_sf_nz_livestock
}
\keyword{datasets}
