% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_sf_nz_river_wq}
\alias{example_sf_nz_river_wq}
\title{Example sf object of New Zealand river water quality trends.}
\format{An \code{sf} object.}
\usage{
example_sf_nz_river_wq
}
\description{
Example sf object of New Zealand river water quality trends.
}
\examples{
example_sf_nz_river_wq
}
\keyword{datasets}
