% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_sf_point}
\alias{example_sf_point}
\title{Example sf point object.}
\format{An \code{sf} object.}
\usage{
example_sf_point
}
\description{
Example sf point object.
}
\examples{
example_sf_point
}
\keyword{datasets}
