% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_stars}
\alias{example_stars}
\title{Example stars object.}
\format{
A \code{stars} object.
}
\usage{
example_stars
}
\description{
Example stars object.
}
\examples{
library(stars)
gg_stars_col(example_stars, col_var = nitrate, borders = example_borders)
}
\keyword{datasets}
