% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_stars_2}
\alias{example_stars_2}
\title{Example stars object 2.}
\format{An \code{stars} object.}
\usage{
example_stars_2
}
\description{
Example stars object.
}
\examples{
example_stars_2
}
\keyword{datasets}
