% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_stars_nz_drp}
\alias{example_stars_nz_drp}
\title{Example stars object of New Zealand modelled river water DRP concentrations.}
\format{An \code{stars} object.}
\usage{
example_stars_nz_drp
}
\description{
Example stars object of New Zealand modelled river water dissolved reactive phosphorus concentrations.
}
\examples{
example_stars_nz_drp
}
\keyword{datasets}
