% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_stars_nz_no3n}
\alias{example_stars_nz_no3n}
\title{Example stars object of New Zealand modelled river water NO3N concentrations.}
\format{An \code{stars} object.}
\usage{
example_stars_nz_no3n
}
\description{
Example stars object of New Zealand modelled river water nitrate concentrations.
}
\examples{
example_stars_nz_no3n
}
\keyword{datasets}
