% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{gg_theme}
\alias{gg_theme}
\title{Create a simple theme for a graph.}
\usage{
gg_theme(
  family = "",
  title_face = "bold",
  title_pal = "#000000",
  title_size = 11,
  subtitle_face = "plain",
  subtitle_pal = "#000000",
  subtitle_size = 10,
  body_face = "plain",
  body_pal = "#323232",
  body_size = 10,
  axis_pal = "#323232",
  axis_size = 0.3,
  ticks_pal = "#323232",
  ticks_size = 0.3,
  gridlines = "horizontal",
  gridlines_pal = "#D3D3D3",
  gridlines_size = 0.2
)
}
\arguments{
\item{family}{The font for all text to use. Defaults to "".}

\item{title_face}{The face of the title font. Defaults to "bold".}

\item{title_pal}{The colour palette for the title font. Defaults to "#000000".}

\item{title_size}{The size of the title font. Defaults to 11.}

\item{subtitle_face}{The face of the subtitle font. Defaults to "plain".}

\item{subtitle_pal}{The colour palette for the subtitle font. Defaults to "#000000".}

\item{subtitle_size}{The size of the subtitle font. Defaults to 10.}

\item{body_face}{The face of all font other than the title or subtitle. Defaults to "plain".}

\item{body_pal}{The colour palette for all font other than the title or subtitle. Defaults to "#323232".}

\item{body_size}{The size of all font other than the title or subtitle. Defaults to 10.}

\item{axis_pal}{The colour palette for the axis. Defaults to "#323232".}

\item{axis_size}{The size of the axis. Defaults to 0.3.}

\item{ticks_pal}{The colour palette for the ticks. Defaults to "#323232".}

\item{ticks_size}{The size of the ticks. Defaults to 0.3.}

\item{gridlines}{Whether gridlines are "horizontal", "vertical", "both" or "none".}

\item{gridlines_pal}{The colour palette for the vertical major gridlines. Defaults to "#D3D3D3".}

\item{gridlines_size}{The size of the vertical major gridlines. Defaults to 0.2.}
}
\value{
A ggplot theme.
}
\description{
Create a simple theme for a graph.
}
