% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme2.R
\name{gg_theme_grid}
\alias{gg_theme_grid}
\title{Create a simple clean theme. Defaults to having horizontal and vertical major gridlines.}
\usage{
gg_theme_grid(
  family = "",
  title_face = "bold",
  title_pal = "#000000",
  title_size = 11,
  subtitle_face = "plain",
  subtitle_pal = "#000000",
  subtitle_size = 10,
  body_face = "plain",
  body_pal = "#323232",
  body_size = 10,
  axis_pal = "#323232",
  axis_size = 0.3,
  ticks_pal = "#323232",
  ticks_size = 0.3,
  grid_h_major_pal = "#D3D3D3",
  grid_h_major_size = 0.2,
  grid_v_major_pal = "#D3D3D3",
  grid_v_major_size = 0.2,
  grid_h_minor_pal = "#D3D3D3",
  grid_h_minor_size = 0,
  grid_v_minor_pal = "#D3D3D3",
  grid_v_minor_size = 0
)
}
\arguments{
\item{family}{The font for all text to use. Defaults to "".}

\item{title_face}{The face of the title font. Defaults to "bold".}

\item{title_pal}{The colour palette for the title font. Defaults to "#000000".}

\item{title_size}{The size of the title font. Defaults to 11.}

\item{subtitle_face}{The face of the subtitle font. Defaults to "plain".}

\item{subtitle_pal}{The colour palette for the subtitle font. Defaults to "#000000".}

\item{subtitle_size}{The size of the subtitle font. Defaults to 10.}

\item{body_face}{The face of all font other than the title or subtitle. Defaults to "plain".}

\item{body_pal}{The colour palette for all font other than the title or subtitle. Defaults to "#323232".}

\item{body_size}{The size of all font other than the title or subtitle. Defaults to 10.}

\item{axis_pal}{The colour palette for the axis. Defaults to "#323232".}

\item{axis_size}{The size of the axis. Defaults to 0.3.}

\item{ticks_pal}{The colour palette for the ticks. Defaults to "#323232".}

\item{ticks_size}{The size of the ticks. Defaults to 0.3.}

\item{grid_h_major_pal}{The colour palette for the horizontal major gridlines. Defaults to "#D3D3D3".}

\item{grid_h_major_size}{The size of the horizontal major gridlines. Defaults to 0.2.}

\item{grid_v_major_pal}{The colour palette for the vertical major gridlines. Defaults to "#D3D3D3".}

\item{grid_v_major_size}{The size of the vertical major gridlines. Defaults to 0.2.}

\item{grid_h_minor_pal}{The colour palette for the horizontal minor gridlines. Defaults to "#D3D3D3".}

\item{grid_h_minor_size}{The size of the horizontal minor gridlines. Defaults to 0.}

\item{grid_v_minor_pal}{The colour palette for the horizontal minor gridlines. Defaults to "#D3D3D3".}

\item{grid_v_minor_size}{The size of the horizontal minor gridlines. Defaults to 0.}
}
\value{
A ggplot theme.
}
\description{
Create a simple clean theme. Defaults to having horizontal and vertical major gridlines.
}
