% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{gg_theme_map}
\alias{gg_theme_map}
\title{Create a simple theme for a map.}
\usage{
gg_theme_map(
  family = "",
  title_face = "bold",
  title_pal = "#000000",
  title_size = 11,
  subtitle_face = "plain",
  subtitle_pal = "#000000",
  subtitle_size = 10,
  body_face = "plain",
  body_pal = "#323232",
  body_size = 10
)
}
\arguments{
\item{family}{The font for all text to use. Defaults to "".}

\item{title_face}{The face of the title font. Defaults to "bold".}

\item{title_pal}{The colour palette for the title font. Defaults to "#000000".}

\item{title_size}{The size of the title font. Defaults to 11.}

\item{subtitle_face}{The face of the subtitle font. Defaults to "plain".}

\item{subtitle_pal}{The colour palette for the subtitle font. Defaults to "#000000".}

\item{subtitle_size}{The size of the subtitle font. Defaults to 10.}

\item{body_face}{The face of all font other than the title or subtitle. Defaults to "plain".}

\item{body_pal}{The colour palette for all font other than the title or subtitle. Defaults to "#323232".}

\item{body_size}{The size of all font other than the title or subtitle. Defaults to 10.}
}
\value{
A ggplot theme.
}
\description{
Create a simple theme for a map.
}
