% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_theme.R
\name{gg_theme_void}
\alias{gg_theme_void}
\title{Create a simple void theme that is useful for maps among other things.}
\usage{
gg_theme_void(
  font = "",
  size_title = 11,
  size_subtitle = 10,
  size_body = 10,
  style_title = "bold",
  style_subtitle = "plain",
  style_body = "plain",
  pal_title = "#000000",
  pal_subtitle = "#000000",
  pal_body = "#323232",
  pal_background = c("#ffffff", "#ffffff")
)
}
\arguments{
\item{font}{The font for all text to use. Defaults to "".}

\item{size_title}{The size of the title font. Defaults to 11.}

\item{size_subtitle}{The size of the subtitle font. Defaults to 10.}

\item{size_body}{The size of all font other than the title or subtitle. Defaults to 10.}

\item{style_title}{The style of the title font. Defaults to "bold".}

\item{style_subtitle}{The style of the subtitle font. Defaults to "plain".}

\item{style_body}{The style of all font other than the title or subtitle. Defaults to "plain".}

\item{pal_title}{The colour palette for the title font. Defaults to "#000000".}

\item{pal_subtitle}{The colour palette for the subtitle font. Defaults to "#000000".}

\item{pal_body}{The colour palette for all font other than the title or subtitle. Defaults to "#323232".}

\item{pal_background}{A two colour vector. The first colour if for the panel (and legend key). The second colour is for the rest of the background.}
}
\value{
A ggplot theme.
}
\description{
Create a simple void theme that is useful for maps among other things.
}
