% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_violin.R
\name{gg_violin_facet}
\alias{gg_violin_facet}
\title{Violin ggplot that is facetted.}
\usage{
gg_violin_facet(
  data,
  x_var,
  y_var = NULL,
  facet_var,
  pal = pal_viridis_reorder(1),
  alpha_fill = 0.2,
  alpha_line = 1,
  size_line = 0.5,
  size_width = 0.75,
  title = NULL,
  title_wrap = 80,
  subtitle = NULL,
  subtitle_wrap = 80,
  x_expand = ggplot2::waiver(),
  x_labels = snakecase::to_sentence_case,
  x_na_rm = FALSE,
  x_rev = FALSE,
  x_title = NULL,
  x_title_wrap = 50,
  y_balance = FALSE,
  y_breaks_n = 3,
  y_expand = c(0, 0),
  y_labels = scales::label_comma(),
  y_title = NULL,
  y_title_wrap = 50,
  y_zero = FALSE,
  y_zero_line = NULL,
  facet_labels = snakecase::to_sentence_case,
  facet_na_rm = FALSE,
  facet_ncol = NULL,
  facet_nrow = NULL,
  facet_rev = FALSE,
  facet_scales = "fixed",
  caption = NULL,
  caption_wrap = 80,
  theme = gg_theme(),
  model_scale = "area",
  model_bw = "nrd0",
  model_adjust = 1,
  model_kernel = "gaussian",
  model_trim = TRUE
)
}
\arguments{
\item{data}{An tibble or dataframe. Required input.}

\item{x_var}{Unquoted categorical variable to be on the x scale (i.e. character, factor, logical). Required input.}

\item{y_var}{Generally an unquoted numeric variable to be on the y scale.}

\item{facet_var}{Unquoted categorical variable to facet the data by. Required input.}

\item{pal}{Character vector of hex codes.}

\item{alpha_fill}{The opacity of the fill. Defaults to 0.2.}

\item{alpha_line}{The opacity of the outline. Defaults to 1.}

\item{size_line}{The size of the outlines of violins. Defaults to 0.5.}

\item{size_width}{Width of boxes. Defaults to 0.75.}

\item{title}{Title string.}

\item{title_wrap}{Number of characters to wrap the title to. Defaults to 75.}

\item{subtitle}{Subtitle string.}

\item{subtitle_wrap}{Number of characters to wrap the subtitle to. Defaults to 75.}

\item{x_expand}{A vector of range expansion constants used to add padding to the x scale, as per the ggplot2 expand argument in ggplot2 scales functions.}

\item{x_labels}{A function or named vector to modify x scale labels. If NULL, categorical variable labels are converted to sentence case. Use ggplot2::waiver() to keep x labels untransformed.}

\item{x_na_rm}{TRUE or FALSE of whether to include x_var NA values. Defaults to FALSE.}

\item{x_rev}{For a categorical x variable, TRUE or FALSE of whether the x variable variable is reversed. Defaults to FALSE.}

\item{x_title}{X scale title string. Defaults to NULL, which converts to sentence case with spaces. Use "" if you would like no title.}

\item{x_title_wrap}{Number of characters to wrap the x title to. Defaults to 50.}

\item{y_balance}{For a numeric y variable, add balance to the y scale so that zero is in the centre of the y scale.}

\item{y_breaks_n}{For a numeric or date x variable, the desired number of intervals on the x scale, as calculated by the pretty algorithm. Defaults to 4.}

\item{y_expand}{A vector of range expansion constants used to add padding to the y scale, as per the ggplot2 expand argument in ggplot2 scales functions.}

\item{y_labels}{A function or named vector to modify y scale labels. Use ggplot2::waiver() to keep y labels untransformed.}

\item{y_title}{y scale title string. Defaults to NULL, which converts to sentence case with spaces. Use "" if you would like no title.}

\item{y_title_wrap}{Number of characters to wrap the y title to. Defaults to 50.}

\item{y_zero}{For a numeric y variable, TRUE or FALSE of whether the minimum of the y scale is zero. Defaults to TRUE.}

\item{y_zero_line}{For a numeric y variable, TRUE or FALSE whether to add a zero reference line to the y scale. Defaults to TRUE if there are positive and negative values in y_var. Otherwise defaults to FALSE.}

\item{facet_labels}{A function or named vector to modify facet scale labels. Defaults to converting labels to sentence case. Use ggplot2::waiver() to keep facet labels untransformed.}

\item{facet_na_rm}{TRUE or FALSE of whether to include facet_var NA values. Defaults to FALSE.}

\item{facet_ncol}{The number of columns of facetted plots.}

\item{facet_nrow}{The number of rows of facetted plots.}

\item{facet_rev}{TRUE or FALSE of whether the facet variable variable is reversed. Defaults to FALSE.}

\item{facet_scales}{Whether facet_scales should be "fixed" across facets, "free" in both directions, or free in just one direction (i.e. "free_x" or "free_y"). Defaults to "fixed".}

\item{caption}{Caption title string.}

\item{caption_wrap}{Number of characters to wrap the caption to. Defaults to 80. #'}

\item{theme}{A ggplot2 theme.}

\item{model_scale}{Per ggplot2::geom_violin, if "area" (default), all violins have the same area (before trimming the tails). If "count", areas are scaled proportionally to the number of observations. If "width", all violins have the same maximum width.}

\item{model_bw}{The bw argument of the stats::density function. Defaults to "nrd0".}

\item{model_adjust}{The adjust argument of the stats::density function. Defaults to 1.}

\item{model_kernel}{The kernel argument of the stats::density function. Defaults to "gaussian".}

\item{model_trim}{TRUE or FALSE of whether to trim the tails. Defaults to FALSE.}
}
\value{
A ggplot object.
}
\description{
Violin ggplot that is facetted, but not coloured.
}
\examples{
library(dplyr)
library(simplevis)
library(palmerpenguins)

gg_violin_facet(penguins, 
                x_var = sex, 
                y_var = body_mass_g, 
                facet_var = species, 
                x_na_rm = TRUE)

}
