% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_hbar.R
\name{ggplot_hbar_facet}
\alias{ggplot_hbar_facet}
\title{Horizontal bar ggplot that is facetted.}
\usage{
ggplot_hbar_facet(data, x_var, y_var, facet_var, text_var = NULL,
  pal = NULL, width = 0.75, alpha = 1, size_line = 0.5,
  title = "[Title]", title_wrap = 70, subtitle = NULL,
  subtitle_wrap = 80, x_balance = FALSE, x_expand = NULL,
  x_labels = waiver(), x_pretty_n = 5, x_title = "[X title]",
  x_title_wrap = 50, x_trans = "identity", x_zero = TRUE,
  x_zero_line = NULL, y_rev = FALSE, y_labels = waiver(),
  y_expand = NULL, y_title = "[Y title]", y_title_wrap = 50,
  facet_scales = "fixed", facet_ncol = NULL, facet_nrow = NULL,
  caption = NULL, font_family = "Helvetica", font_size_title = NULL,
  font_size_body = NULL, caption_wrap = 80)
}
\arguments{
\item{data}{A tibble or dataframe. Required input.}

\item{x_var}{Unquoted numeric variable to be on the x axis. Required input.}

\item{y_var}{Unquoted categorical variable to be on the y axis. Required input.}

\item{facet_var}{Unquoted categorical variable to facet the data by. Required input.}

\item{text_var}{Unquoted variable to be used as a customised tooltip in combination with plotly::ggplotly(plot, tooltip = "text"). Defaults to NULL.}

\item{pal}{Character vector of hex codes. Defaults to viridis. Use the pals package to find a suitable palette.}

\item{width}{Width of bars. Defaults to 0.75.}

\item{alpha}{The alpha of the fill. Defaults to 1.}

\item{size_line}{The size of the outlines of bars.}

\item{title}{Title string. Defaults to [Title].}

\item{title_wrap}{Number of characters to wrap the title to. Defaults to 70.}

\item{subtitle}{Subtitle string. Defaults to [Subtitle].}

\item{subtitle_wrap}{Number of characters to wrap the subtitle to. Defaults to 80.}

\item{x_balance}{Add balance to the x axis so that zero is in the centre of the x scale. Only applicable where facet_scales equals "fixed" or "free_y".}

\item{x_expand}{A vector of range expansion constants used to add some padding on the x scale.}

\item{x_labels}{Adjust the  x scale labels through a function or vector.}

\item{x_pretty_n}{The desired number of intervals on the x axis, as calculated by the pretty algorithm. Defaults to 5.}

\item{x_title}{X axis title string. Defaults to [X title].}

\item{x_title_wrap}{Number of characters to wrap the x title to. Defaults to 50.}

\item{x_trans}{A string specifying a transformation for the x scale. Defaults to "identity".}

\item{x_zero}{TRUE or FALSE whether the minimum of the x scale is zero. Defaults to TRUE.}

\item{x_zero_line}{TRUE or FALSE whether to add a zero reference line to the x axis. Defaults to NULL, which is TRUE if there are positive and negative values in x_var. Otherwise it is FALSE.}

\item{y_rev}{TRUE or FALSE of whether bar order from top to bottom is reversed from default. Defaults to FALSE.}

\item{y_labels}{Adjust the  y scale labels through a function or vector.}

\item{y_expand}{A vector of range expansion constants used to add some padding on the y scale.}

\item{y_title}{Y axis title string. Defaults to [Y title].}

\item{y_title_wrap}{Number of characters to wrap the y title to. Defaults to 50.}

\item{facet_scales}{Whether facet_scales should be "fixed" across facets, "free" in both directions, or free in just one direction (i.e. "free_x" or "free_y"). Defaults to "fixed".}

\item{facet_ncol}{The number of columns of facetted plots.}

\item{facet_nrow}{The number of rows of facetted plots.}

\item{caption}{Caption title string. Defaults to NULL.}

\item{font_family}{Font family to use. Defaults NULL.}

\item{font_size_title}{Font size for the title text. Defaults to 11.}

\item{font_size_body}{Font size for all text other than the title. Defaults to 10.}

\item{caption_wrap}{Number of characters to wrap the caption to. Defaults to 80.}
}
\value{
A ggplot object.
}
\description{
Horizontal bar ggplot that is facetted, but not coloured.
}
\examples{
library(dplyr)

plot_data <- ggplot2::diamonds \%>\%
  mutate(cut = stringr::str_to_sentence(cut)) \%>\%
  group_by(cut, clarity) \%>\%
  summarise(average_price = mean(price)) \%>\%
  mutate(average_price = round(average_price / 1000, 1)) 

ggplot_hbar_facet(plot_data, average_price, cut, clarity,
   title = "Average diamond price by cut and clarity", 
   x_title = "Average price ($US thousands)", 
   y_title = "Cut")

}
