% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{interval_labels}
\alias{interval_labels}
\title{Convert bin cuts to interval legend labels.}
\usage{
interval_labels(cuts, digits = NULL, right_closed = TRUE)
}
\arguments{
\item{cuts}{A numeric vector of bin cuts from which to create a vector of legend labels.}

\item{digits}{The number of decimal places to round labels to.}

\item{right_closed}{TRUE or FALSE of whether bins or quantiles are to be cut right-closed. Defaults to TRUE.}
}
\value{
A vector of labels.
}
\description{
Convert bin cuts to interval legend labels.
}
\examples{
interval_labels(c(0, 0.1, 3, 4.1, 7, 100, Inf))
}
