% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaf_basemap.R
\name{leaf_basemap}
\alias{leaf_basemap}
\title{Basemap stack in leaflet.}
\usage{
leaf_basemap(bounds = NULL, basemap = "light")
}
\arguments{
\item{bounds}{A bbox object or numeric vector of length four, with xmin, ymin, xmax and ymax values in WGS84 (epsg 4326).}

\item{basemap}{The first layer to start in the basemap stack. Either "light", "dark", "street", "satellite", or "ocean". Defaults to "light".}
}
\value{
A leaflet object.
}
\description{
Make a stack of leaflet baselayers for use in shiny apps.
}
\examples{
leaf_basemap(basemap = "dark")

leaf_basemap(bounds = c(166.70047,-34.45676, 178.52966,-47.06345))

}
