% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_basemap_fun.R
\name{leaflet_basemap}
\alias{leaflet_basemap}
\title{Basemap stack in leaflet.}
\usage{
leaflet_basemap(top_layer = "light")
}
\arguments{
\item{top_layer}{The first layer to start in the basemap stack. Either "light", "dark", "street", "satellite", or "ocean". Defaults to "light".}
}
\value{
A leaflet object.
}
\description{
Make a stack of leaflet baselayers for use in shiny apps.
}
\examples{
leaflet_basemap("dark")

bb <- rnaturalearth::ne_countries(scale = "small", country = "Indonesia", returnclass = "sf") \%>\% 
sf::st_bbox() \%>\% 
  as.vector()

simplevis::leaflet_basemap("satellite") \%>\% 
  leaflet::fitBounds(bb[1], bb[2], bb[3], bb[4])
  
}
