% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_basemap_fun.R
\name{leaflet_basemap_nz}
\alias{leaflet_basemap_nz}
\title{Basemap stack in leaflet for New Zealand.}
\usage{
leaflet_basemap_nz(top_layer = "light")
}
\arguments{
\item{top_layer}{The first layer to start in the basemap stack. Either "light", "dark", "street", "satellite", or "ocean". Defaults to "light".}
}
\value{
A leaflet object.
}
\description{
Make a stack of leaflet baselayers for use in New Zealand focussed shiny apps.
}
\examples{
leaflet_basemap_nz("dark")
}
