% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_sf.R
\name{leaflet_sf}
\alias{leaflet_sf}
\title{Simple feature leaflet map.}
\usage{
leaflet_sf(
  data,
  popup_vars_vctr = NULL,
  popup_vars_rename = snakecase::to_sentence_case,
  pal = pal_viridis_reorder(1),
  size_point = 2,
  size_line = 2,
  alpha_point = NULL,
  alpha_line = NULL,
  alpha_fill = NULL,
  basemap = "light",
  map_id = "map"
)
}
\arguments{
\item{data}{An sf object of geometry type point/multipoint, linestring/multilinestring or polygon/multipolygon geometry type. Required input.}

\item{popup_vars_vctr}{Vector of quoted variable names to include in the popup. If NULL, defaults to making a leafpop::popupTable of all columns.}

\item{popup_vars_rename}{Function to rename column names for the popup. Defaults to snakecase::to_sentence_case. Use function(x) x to leave column names untransformed.}

\item{pal}{Character vector of hex codes.}

\item{size_point}{Size of points (i.e. radius). Defaults to 2.}

\item{size_line}{Size of lines around features (i.e. weight). Defaults to 2.}

\item{alpha_point}{The opacity of the points.}

\item{alpha_line}{The opacity of the outline.}

\item{alpha_fill}{The opacity of the fill.}

\item{basemap}{The underlying basemap. Either "light", "dark", "satellite", "street", or "ocean". Defaults to "light". Only applicable where shiny equals FALSE.}

\item{map_id}{The shiny map id for a leaflet map within a shiny app. For standard single-map apps, id "map" should be used. For dual-map apps, "map1" and "map2" should be used. Defaults to "map".}
}
\value{
A leaflet object.
}
\description{
Map of simple features in leaflet that is not coloured.
}
\examples{
leaflet_sf(example_sf_point)

leaflet_sf(example_sf_polygon)

}
