% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_stars.R
\name{leaflet_stars_col}
\alias{leaflet_stars_col}
\title{Stars leaflet map that is coloured.}
\usage{
leaflet_stars_col(data, col_var, band = 1, pal = NULL,
  pal_na = "#7F7F7F", pal_rev = FALSE, alpha = 1,
  basemap = "light", title = NULL, col_cuts = NULL,
  col_labels = NULL, col_method = NULL, col_pretty_n = 4,
  col_right_closed = TRUE, map_id = "map")
}
\arguments{
\item{data}{A stars object. Required input.}

\item{col_var}{Unquoted attribute to colour the features by. Required input.}

\item{band}{The band number to be plotted.}

\item{pal}{Character vector of hex codes.}

\item{pal_na}{The hex code or name of the NA colour to be used.}

\item{pal_rev}{Reverses the palette. Defaults to FALSE.}

\item{alpha}{The opacity of features. Defaults to 1.}

\item{basemap}{The underlying basemap. Either "light", "dark", "satellite", "street", or "ocean". Defaults to "light". Only applicable where shiny equals FALSE.}

\item{title}{A title string that will be wrapped into the legend.}

\item{col_cuts}{A vector of cuts to colour a numeric variable. If "bin" is selected, the first number in the vector should be either -Inf or 0, and the final number Inf. If "quantile" is selected, the first number in the vector should be 0 and the final number should be 1. Defaults to quartiles.}

\item{col_labels}{A vector to modify colour scale labels.}

\item{col_method}{The method of colouring features, either "bin", "quantile" or "category." If numeric, defaults to "bin".}

\item{col_pretty_n}{For a numeric colour variable of "bin" col_method, the desired number of intervals on the colour scale, as calculated by the pretty algorithm. Defaults to 4.}

\item{col_right_closed}{For a numeric colour variable, TRUE or FALSE of whether bins or quantiles are to be cut right-closed. Defaults to TRUE.}

\item{map_id}{The shiny map id for a leaflet map within a shiny app. For standard single-map apps, id "map" should be used. For dual-map apps, "map1" and "map2" should be used. Defaults to "map".}
}
\value{
A leaflet object.
}
\description{
Map of stars in leaflet that is coloured.
}
\examples{
library(simplevis)
library(stars)

leaflet_stars_col(example_stars, 
                  col_var = NO3N_p50.tif, 
                  pal_na = "transparent")

leaflet_stars_col(example_stars, 
                  col_var = NO3N_p50.tif, 
                  col_method = "bin", 
                  col_cuts = c(seq(0, 3000, 500), Inf),
                  pal_na = "transparent") 

leaflet_stars_col(example_stars, 
                  col_var = NO3N_p50.tif, 
                  col_method = "bin", 
                  col_pretty_n = 7,
                  pal_na = "transparent")

cuts <- quantile(example_stars$NO3N_p50.tif,
                 probs = c(0, 0.05, 0.25, 0.5, 0.75, 0.95, 1), na.rm = TRUE)

labels <- interval_labels(cuts = cuts, digits = 2)

leaflet_stars_col(example_stars, 
                  col_var = NO3N_p50.tif, 
                  col_method = "quantile", 
                  col_cuts = c(0, 0.05, 0.25, 0.5, 0.75, 0.95, 1), 
                  col_labels = labels,
                  pal_na = "transparent") 

}
