% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{legend_labels_from_cuts}
\alias{legend_labels_from_cuts}
\title{Numeric legend labels.}
\usage{
legend_labels_from_cuts(cuts_vctr, col_digits = 1)
}
\arguments{
\item{cuts_vctr}{A numeric vector of bin cuts from which to create a vector of legend labels.}

\item{col_digits}{The number of digits to round the legend labels.}
}
\value{
A vector of labels.
}
\description{
Pretty numeric legend labels.
}
\keyword{internal}
