% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natural_earth_fun.R
\name{ne_boundary}
\alias{ne_boundary}
\title{ne_boundary}
\usage{
ne_boundary(country = NULL, state = NULL, continent = NULL,
  scale = "large", crs = 4326)
}
\arguments{
\item{country}{A country name from Natural Earth.}

\item{state}{A state or region name within a country from Natural Earth. Must be used in combination with the applicable country specified.}

\item{continent}{A continent name from Natural Earth.}

\item{scale}{Scale of map to extract the bounds from with "small", "medium" or "large" options. Defaults to "large". Not applicable to where a state is extracted.}

\item{crs}{Coordinate reference system of bounds. Defaults to 4326.}
}
\value{
An sf object.
}
\description{
Extract a boundary from Natural Earth.
}
\examples{
ggplot_sf(ne_boundary("Papua New Guinea"))
ggplot_sf(ne_boundary("Australia", state = TRUE))
ggplot_sf(ne_boundary("Australia", state = "Tasmania"))
}
