% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natural_earth_fun.R
\name{ne_bounds}
\alias{ne_bounds}
\title{ne_bounds}
\usage{
ne_bounds(country = NULL, state = NULL, continent = NULL,
  scale = "small", crs = 4326)
}
\arguments{
\item{country}{A country name from Natural Earth.}

\item{state}{A state (or region) name within a country from Natural Earth or TRUE to specify all state boundaries. Must be used in combination with the applicable country specified.}

\item{continent}{A continent name from Natural Earth.}

\item{scale}{Scale of map to extract the bounds from with "small", "medium" or "large" options. Defaults to "small".}

\item{crs}{Coordinate reference system of bounds. Defaults to 4326.}
}
\value{
A vector of the bounding box.
}
\description{
Extract a vector of bounds from Natural Earth
}
\examples{
ne_bounds("Papua New Guinea")
ne_bounds("Australia", state = TRUE)
ne_bounds("Australia", state = "Tasmania")
}
