% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{numeric_legend_labels}
\alias{numeric_legend_labels}
\title{Numeric legend labels.}
\usage{
numeric_legend_labels(bin_cuts, legend_digits = 1)
}
\arguments{
\item{bin_cuts}{A numeric vector of bin cuts from which to create a vector of legend labels.}

\item{legend_digits}{The number of digits to round the legend labels.}
}
\value{
A vector of labels.
}
\description{
Pretty numeric legend labels.
}
