% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nz_region}
\alias{nz_region}
\title{New Zealand coastline with regions.}
\format{An \code{sf} object.}
\source{
\url{https://datafinder.stats.govt.nz/layer/98765-regional-council-2019-clipped-generalised/}
}
\usage{
nz_region
}
\description{
New Zealand coastline, excluding the Chathams, with regional boundaries simplified for ggplot.
}
\examples{
nz_region

ggplot_sf(nz_region)

ggplot_sf(dplyr::filter(nz_region, region_code == "09"))

}
\keyword{datasets}
