% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{pal_default}
\alias{pal_default}
\title{Get default palette.}
\usage{
pal_default(n_col)
}
\arguments{
\item{data}{The number of digits to round the legend labels.}
}
\value{
A numeric value.
}
\description{
A function to for making hover values.
}
\keyword{internal}
