% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/er_helpers.R
\docType{data}
\name{pal_ea19}
\alias{pal_ea19}
\title{Colour palette for categorical variables.}
\format{An object of class \code{character} of length 9.}
\usage{
pal_ea19
}
\value{
A vector of hex codes.
}
\description{
Colour palette for categorical variables.
}
\keyword{datasets}
