% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\docType{data}
\name{pal_ea19_nztcs_subcategory_change}
\alias{pal_ea19_nztcs_subcategory_change}
\title{pal_ea19_nztcs_subcategory_change}
\format{An object of class \code{character} of length 2.}
\usage{
pal_ea19_nztcs_subcategory_change
}
\value{
A vector of hex codes.
}
\description{
A colour palette used for depicting categories in the NZ conservation threat status.
}
\keyword{datasets}
