% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/er_helpers.R
\docType{data}
\name{pal_ea19_trend3}
\alias{pal_ea19_trend3}
\title{Colour palette for a categorical trend variable with 3 values.}
\format{An object of class \code{character} of length 3.}
\usage{
pal_ea19_trend3
}
\value{
A vector of hex codes.
}
\description{
Colour palette for categorical variables.
}
\examples{
scales::show_col(pal_ea19_trend3)
}
\keyword{datasets}
