% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/er_helpers.R
\docType{data}
\name{pal_point_trend5}
\alias{pal_point_trend5}
\title{Colour palette for a categorical trend variable with 5 values for points.}
\format{An object of class \code{character} of length 5.}
\usage{
pal_point_trend5
}
\value{
A vector of hex codes.
}
\description{
Colour palette for categorical variables.
}
\examples{
scales::show_col(pal_point_trend5)
}
\keyword{datasets}
