% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/er_helpers.R
\docType{data}
\name{pal_snz_compare2}
\alias{pal_snz_compare2}
\title{Colour palette for a graph that compares a current year to 1 past year.}
\format{An object of class \code{character} of length 2.}
\usage{
pal_snz_compare2
}
\value{
A vector of hex codes. Uses the first colour, and then a 30% tint.
}
\description{
Colour palette for a graph that compares a current year to 1 past year.
}
\examples{
scales::show_col(pal_snz_compare2)
}
\keyword{datasets}
