% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/er_helpers.R
\docType{data}
\name{pal_snz_nof4}
\alias{pal_snz_nof4}
\title{Colour palette for 4 categories from good to bad}
\format{An object of class \code{character} of length 4.}
\usage{
pal_snz_nof4
}
\value{
A vector of hex codes.
}
\description{
A colour palette used for depicting subcategories in the NZ conservation threat status.
}
\keyword{datasets}
