% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/er_helpers.R
\docType{data}
\name{pal_snz_nof5}
\alias{pal_snz_nof5}
\title{Colour palette for 5 categories from good to bad}
\format{An object of class \code{character} of length 5.}
\usage{
pal_snz_nof5
}
\value{
A vector of hex codes.
}
\description{
A colour palette used for depicting subcategories in the NZ conservation threat status.
}
\examples{
scales::show_col(pal_snz_nof5)
}
\keyword{datasets}
