% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\docType{data}
\name{pal_snz_nztcs_category}
\alias{pal_snz_nztcs_category}
\title{pal_snz_nztcs_category}
\format{An object of class \code{character} of length 4.}
\usage{
pal_snz_nztcs_category
}
\value{
A vector of hex codes.
}
\description{
A colour palette used for depicting categories in the NZ conservation threat status.
}
\keyword{datasets}
