% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/er_helpers.R
\docType{data}
\name{pal_snz_nztcs_s}
\alias{pal_snz_nztcs_s}
\title{Colour palette for the NZTCS.}
\format{An object of class \code{character} of length 9.}
\usage{
pal_snz_nztcs_s
}
\value{
A vector of hex codes.
}
\description{
A colour palette used for depicting subcategories in the NZ conservation threat status.
}
\keyword{datasets}
