% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\docType{data}
\name{pal_snz_nztcs_subcategory}
\alias{pal_snz_nztcs_subcategory}
\title{pal_snz_nztcs_subcategory}
\format{An object of class \code{character} of length 9.}
\usage{
pal_snz_nztcs_subcategory
}
\value{
A vector of hex codes.
}
\description{
A colour palette used for depicting subcategories in the NZ conservation threat status.
}
\keyword{datasets}
