% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/er_helpers.R
\docType{data}
\name{pal_snz_trend2}
\alias{pal_snz_trend2}
\title{Colour palette for a categorical trend variable with 2 values.}
\format{An object of class \code{character} of length 2.}
\usage{
pal_snz_trend2
}
\value{
A vector of hex codes.
}
\description{
Colour palette for categorical variables.
}
\examples{
scales::show_col(pal_snz_trend2)
}
\keyword{datasets}
