% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/er_helpers.R
\docType{data}
\name{pal_snz_trend5}
\alias{pal_snz_trend5}
\title{Colour palette for a categorical trend variable with 5 values.}
\format{An object of class \code{character} of length 5.}
\usage{
pal_snz_trend5
}
\value{
A vector of hex codes.
}
\description{
Colour palette for categorical variables.
}
\keyword{datasets}
