% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\docType{data}
\name{pal_trend3}
\alias{pal_trend3}
\title{Colour palette for categorical variables.}
\format{An object of class \code{character} of length 3.}
\usage{
pal_trend3
}
\value{
A vector of hex codes.
}
\description{
Colour palette for categorical variables.
}
\keyword{datasets}
