% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.R
\name{pal_viridis_reorder}
\alias{pal_viridis_reorder}
\title{Viridis palette reordered.}
\usage{
pal_viridis_reorder(n_col)
}
\arguments{
\item{n_col}{The number of colours (excluding an NA colour).}
}
\value{
A vector of hex codes.
}
\description{
A function to retreive a vector of hex codes for a numeric (or ordererd) variable.
}
