% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly.R
\name{plotly_remove_buttons}
\alias{plotly_remove_buttons}
\title{Remove plotly buttons from the mode bar, other than the camera and plotly logo.}
\usage{
plotly_remove_buttons(plotly, logo = FALSE)
}
\arguments{
\item{plotly}{A plotly object.}

\item{logo}{TRUE or FALSE of whether to display the plotly logo. Defaults to FALSE.}
}
\description{
Remove plotly buttons from the mode bar, other than the camera and plotly logo.
}
\examples{
plot_data <- dplyr::sample_frac(ggplot2::diamonds, 0.05)

plot <- ggplot_scatter(data = plot_data, x_var = carat, y_var = price)

plotly::ggplotly(plot, tooltip = "text") \%>\% 
   plotly_remove_buttons()
}
