% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{remove_plotly_buttons}
\alias{remove_plotly_buttons}
\title{Remove ggplotly buttons from the mode bar, other than the camera and plotly logo.}
\usage{
remove_plotly_buttons(plotly, logo = FALSE)
}
\arguments{
\item{plotly}{A plotly object.}

\item{logo}{TRUE or FALSE of whether to display the plotly logo. Defaults to FALSE.}
}
\description{
Remove ggplotly buttons from the mode bar, other than the camera and plotly logo.
}
