% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app_fun.R
\name{run_example}
\alias{run_example}
\title{Run shiny example with option to download.}
\usage{
run_example(example = "example1", ...)
}
\arguments{
\item{example}{example name. Available examples are "example1" for a graph and table, and "example2" and "example3" also providing maps. Defaults to "example1".}

\item{...}{passed to \code{shiny::runApp}}
}
\description{
Run shiny example with option to download.
}
