% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app_fun.R
\name{run_template}
\alias{run_template}
\title{Run shiny template with option to download.}
\usage{
run_template(template = "template1", ...)
}
\arguments{
\item{template}{template name. Available templates are "template1" for a graph and table, and "template2" and "template3" also providing maps. Defaults to "template1".}

\item{...}{passed to \code{shiny::runApp}}
}
\description{
Run shiny template with option to download.
}
