% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{sentence_colnames}
\alias{sentence_colnames}
\title{Convert column names to sentence case}
\usage{
sentence_colnames(data)
}
\arguments{
\item{data}{The number of digits to round the legend labels.}
}
\value{
A numeric value.
}
\description{
A function to convert colnames to snakecase and then to sentence case to be used in functions for making hover values.
}
