% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_demo.R
\name{shiny_demo}
\alias{shiny_demo}
\title{Shiny demo}
\usage{
shiny_demo(mobile = FALSE)
}
\arguments{
\item{mobile}{TRUE or FALSE of whether the app and code should also work on mobile devices. Defaults to FALSE.}
}
\description{
Run a shiny demo app with option to download code.
}
