% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_template.R
\name{shiny_template}
\alias{shiny_template}
\title{shiny_template}
\usage{
shiny_template(template = "template1", ...)
}
\arguments{
\item{template}{template name. Available templates are "template1" with graph and table tabs, and "template2" also with a map tab. Defaults to "template1".}

\item{...}{passed to \code{shiny::runApp}}
}
\description{
Run a shiny template with option to download.
}
