% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/er_helpers.R
\name{signed_sqrt_trans}
\alias{signed_sqrt_trans}
\title{Signed square root ggplot scale transformation.}
\usage{
signed_sqrt_trans()
}
\value{
A ggplot scale transformation.
}
\description{
A signed square root ggplot scale transformation.
}
