% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplevis-package.R
\docType{package}
\name{simplevis-package}
\alias{simplevis}
\alias{simplevis-package}
\title{simplevis: Simple 'ggplot2' and 'leaflet' Visualisation with Less Brainpower}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Simple 'ggplot2' and 'leaflet' Visualisation with Less Brainpower.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://StatisticsNZ.github.io/simplevis/}
  \item \url{https://github.com/StatisticsNZ/simplevis/}
  \item Report bugs at \url{https://github.com/StatisticsNZ/simplevis/issues/}
}

}
\author{
\strong{Maintainer}: David Hodge \email{davidhodge931@gmail.com} (\href{https://orcid.org/0000-0002-3868-7501}{ORCID})

Other contributors:
\itemize{
  \item Kate Lee (\href{https://orcid.org/0000-0002-0886-3746}{ORCID}) [contributor]
  \item Xavier Miles (\href{https://orcid.org/0000-0002-1727-5110}{ORCID}) [contributor]
  \item Statistics New Zealand [copyright holder]
}

}
\keyword{internal}
