% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\docType{data}
\name{spectrum}
\alias{spectrum}
\title{A colour palette for nominal variables/}
\format{
An object of class \code{character} of length 12.
}
\usage{
spectrum
}
\arguments{
\item{n_col}{The number of col_var levels or values.}
}
\value{
A character vector.
}
\description{
A colour palette for nominal variables/
}
\keyword{internal}
