% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{sv_colnames_to_present}
\alias{sv_colnames_to_present}
\title{Convert column names to snake case and then to sentence case.}
\usage{
sv_colnames_to_present(data)
}
\arguments{
\item{data}{The number of digits to round the legend labels.}
}
\value{
A numeric value.
}
\description{
Convert colnames to snakecase and then to sentence case to present colnames visually.
}
\keyword{internal}
