% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{sv_cuts_to_interval_labels}
\alias{sv_cuts_to_interval_labels}
\title{Convert cuts to bin legend labels.}
\usage{
sv_cuts_to_interval_labels(cuts_vctr, labels_dp = NULL, right_closed = TRUE)
}
\arguments{
\item{cuts_vctr}{A numeric vector of bin cuts from which to create a vector of legend labels.}

\item{labels_dp}{The number of decimal places to round labels to.}

\item{right_closed}{TRUE or FALSE of whether intervals are to be cut right-closed. Defaults to TRUE.}
}
\value{
A vector of labels.
}
\description{
Convert cuts to bin legend labels.
}
\keyword{internal}
\keyword{sv_cuts_to_labels}
