% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{sv_cuts_to_labels}
\alias{sv_cuts_to_labels}
\title{Convert cuts to bin legend labels.}
\usage{
sv_cuts_to_labels(cuts_vctr, labels_dp = NULL, comma = TRUE)
}
\arguments{
\item{cuts_vctr}{A numeric vector of bin cuts from which to create a vector of legend labels.}

\item{labels_dp}{The number of decimal places to round numeric labels to.}

\item{comma}{TRUE or FALSE of whether to convert numeric values to character values with comma seperators.}
}
\value{
A vector of labels.
}
\description{
Convert cuts to bin legend labels.
}
\keyword{internal}
