% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_max.R
\name{sv_density_max}
\alias{sv_density_max}
\title{Calculate the max density in a gg_density() plot.}
\usage{
sv_density_max(
  data,
  x_var,
  alg_bw = "nrd0",
  alg_adjust = 1,
  alg_kernel = "gaussian",
  alg_n = 512,
  alg_trim = FALSE
)
}
\arguments{
\item{data}{A tibble or dataframe. Required input.}

\item{x_var}{Unquoted numeric variable to be on the x scale. Required input.}

\item{alg_bw}{The bw argument of the stats::density function. Defaults to "nrd0".}

\item{alg_adjust}{The adjust argument of the stats::density function. Defaults to 1.}

\item{alg_kernel}{The kernel argument of the stats::density function. Defaults to "gaussian".}

\item{alg_n}{The n argument of the stats::density function. Defaults to 512.}

\item{alg_trim}{The trim argument of the stats::density function. Defaults to FALSE.}
}
\description{
Calculate the max density in a gg_density() plot.
}
\keyword{internal}
