% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font_size.R
\name{sv_font_size_body}
\alias{sv_font_size_body}
\title{Get default font_size_body}
\usage{
sv_font_size_body(mobile)
}
\value{
values for font sizes
}
\description{
Get default font_size_body
}
\keyword{internal}
