% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{sv_font_size_title}
\alias{sv_font_size_title}
\title{Get default font_size_title}
\usage{
sv_font_size_title(mobile)
}
\value{
values for font sizes
}
\description{
Get default font_size_title
}
\keyword{internal}
