% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{sv_interval_breaks_to_interval_labels}
\alias{sv_interval_breaks_to_interval_labels}
\title{Convert interval labels to a simpler format.}
\usage{
sv_interval_breaks_to_interval_labels(breaks)
}
\arguments{
\item{breaks}{A vector of interval breaks.}
}
\value{
Labels for the legend.
}
\description{
Convert interval labels to a simpler format.
}
\keyword{internal}
