% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_labels.R
\name{sv_interval_labels_chr}
\alias{sv_interval_labels_chr}
\title{Convert bin cuts to simple and pretty labels.}
\usage{
sv_interval_labels_chr(cuts)
}
\arguments{
\item{cuts}{An vector of interval cuts.}
}
\value{
A vector of character labels.
}
\description{
Convert bin cuts to simple and pretty labels.
}
\examples{
library(simplevis)

interval_labels(c("(0, 10]", "(10, 50]", "(50, 100]"))
}
\keyword{internal}
