% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_labels.R
\name{sv_interval_labels_num}
\alias{sv_interval_labels_num}
\title{Convert numeric bin cuts to simple and pretty labels.}
\usage{
sv_interval_labels_num(cuts, format = NULL, right_closed = TRUE)
}
\arguments{
\item{cuts}{A vector of numeric cuts.}

\item{format}{A function to apply to the numeric values, e.g. scales::label_comma().}

\item{right_closed}{If cuts are numeric, TRUE or FALSE of whether intervals are to be right-closed. Defaults to TRUE.}
}
\value{
A vector of character labels.
}
\description{
Convert numeric bin cuts to simple and pretty labels.
}
\keyword{internal}
