% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_density.R
\name{sv_max_density}
\alias{sv_max_density}
\title{Calculate the max density in a gg_density() plot.}
\usage{
sv_max_density(
  data,
  x_var,
  density_bw = "nrd0",
  density_adjust = 1,
  density_kernel = "gaussian",
  density_n = 512,
  density_trim = FALSE
)
}
\arguments{
\item{data}{A tibble or dataframe. Required input.}

\item{x_var}{Unquoted numeric variable to be on the x scale. Required input.}

\item{density_bw}{The bw argument of the stats::density function. Defaults to "nrd0".}

\item{density_adjust}{The adjust argument of the stats::density function. Defaults to 1.}

\item{density_kernel}{The kernel argument of the stats::density function. Defaults to "gaussian".}

\item{density_n}{The n argument of the stats::density function. Defaults to 512.}

\item{density_trim}{The trim argument of the stats::density function. Defaults to FALSE.}
}
\description{
Calculate the max density in a gg_density() plot.
}
\keyword{internal}
