% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_density.R
\name{sv_max_density_col_stack}
\alias{sv_max_density_col_stack}
\title{Calculate the max density in a gg_density_col(..., position == "stack") plot.}
\usage{
sv_max_density_col_stack(
  data,
  x_var,
  col_var,
  density_bw = "nrd0",
  density_adjust = 1,
  density_kernel = "gaussian",
  density_n = 512,
  density_trim = FALSE
)
}
\arguments{
\item{data}{A tibble or dataframe. Required input.}

\item{x_var}{Unquoted numeric variable to be on the x scale. Required input.}

\item{col_var}{Unquoted categorical variable to colour density areas. Required input.}
}
\description{
Calculate the max density in a gg_density_col(..., position == "stack") plot.
}
\keyword{internal}
