% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_density.R
\name{sv_max_density_facet}
\alias{sv_max_density_facet}
\title{Calculate the max density in a gg_density_facet() plot.}
\usage{
sv_max_density_facet(
  data,
  x_var,
  facet_var,
  density_bw = "nrd0",
  density_adjust = 1,
  density_kernel = "gaussian",
  density_n = 512,
  density_trim = FALSE
)
}
\arguments{
\item{data}{A tibble or dataframe. Required input.}

\item{x_var}{Unquoted numeric variable to be on the x scale. Required input.}

\item{facet_var}{Unquoted categorical variable to facet by. Required input.}
}
\description{
Calculate the max density in a gg_density_facet() plot.
}
\keyword{internal}
