% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_digits.R
\name{sv_max_digits}
\alias{sv_max_digits}
\title{Identify the maximum decimal places in a numeric vector.}
\usage{
sv_max_digits(vctr)
}
\arguments{
\item{vctr}{A numeric vector.}
}
\value{
a numeric value.
}
\description{
Identify the maximum decimal places in a numeric vector.
}
\keyword{internal}
