% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{sv_max_dp}
\alias{sv_max_dp}
\title{Identify the maximum decimal places in a numeric vector.}
\usage{
sv_max_dp(vctr)
}
\arguments{
\item{vctr}{A numeric vector.}
}
\value{
a numeric value.
}
\description{
Identify the maximum decimal places in a numeric vector.
}
\keyword{internal}
