% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{sv_numeric_bin_labels}
\alias{sv_numeric_bin_labels}
\title{Numeric legend labels.}
\usage{
sv_numeric_bin_labels(cuts_vctr, labels_dp = 1)
}
\arguments{
\item{cuts_vctr}{A numeric vector of bin cuts from which to create a vector of legend labels.}

\item{labels_dp}{The number of digits to round the legend labels.}
}
\value{
A vector of labels.
}
\description{
Pretty numeric legend labels.
}
\keyword{internal}
