% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{sv_pal}
\alias{sv_pal}
\title{Get default palette.}
\usage{
sv_pal(n_col)
}
\arguments{
\item{n_col}{The number of col_var levels or values.}
}
\value{
A numeric value.
}
\description{
Get default palette.
}
\keyword{internal}
