% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{sv_x_zero_adjust}
\alias{sv_x_zero_adjust}
\title{Automatically adjust x_zero and x_zero_line if necessary}
\usage{
sv_x_zero_adjust(x_var_vctr, x_balance, x_zero, x_zero_line)
}
\arguments{
\item{x_var_vctr}{A vector of values for the x scale.}

\item{x_balance}{Add balance to the x axis so that zero is in the centre of the x scale. Only applicable where facet_scales equals "fixed" or "free_y".}

\item{x_zero}{TRUE or FALSE whether the minimum of the x scale is zero. Defaults to TRUE.}

\item{x_zero_line}{TRUE or FALSE whether to add a zero reference line to the x axis. Defaults to NULL, which is TRUE if there are positive and negative values in x_var. Otherwise it is FALSE.}

\item{description}{Generate a list of x_zero and x_zero_line elements that are adjusted if necessary.}
}
\value{
A list with first element x_zero (TRUE or FALSE) and second element x_zero_line (TRUE or FALSE)
}
\description{
Automatically adjust x_zero and x_zero_line if necessary
}
\keyword{internal}
