% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{sv_y_zero_adjust}
\alias{sv_y_zero_adjust}
\title{Automatically adjust y_zero and y_zero_line if necessary}
\usage{
sv_y_zero_adjust(y_var_vctr, y_balance, y_zero, y_zero_line)
}
\arguments{
\item{y_var_vctr}{A vector of values for the y scale.}

\item{y_balance}{Add balance to the y axis so that zero is in the centre of the y scale. Only applicable where facet_scales equals "fixed" or "free_y".}

\item{y_zero}{TRUE or FALSE whether the minimum of the y scale is zero. Defaults to TRUE.}

\item{y_zero_line}{TRUE or FALSE whether to add a zero reference line to the y axis. Defaults to NULL, which is TRUE if there are positive and negative values in y_var. Otherwise it is FALSE.}

\item{description}{Generate a list of y_zero and y_zero_line elements that are adjusted if necessary.}
}
\value{
A list with first element y_zero (TRUE or FALSE) and second element y_zero_line (TRUE or FALSE)
}
\description{
Automatically adjust y_zero and y_zero_line if necessary
}
\keyword{internal}
