% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_hbar_fun.R
\name{theme_hbar_mobile_adj}
\alias{theme_hbar_mobile_adj}
\title{Theme mobile adjustment for horizontal bar ggplots.}
\usage{
theme_hbar_mobile_adj(font_family = "Helvetica", font_size_title = 11,
  font_size_body = 10)
}
\arguments{
\item{font_family}{Font family to use. Defaults to "Helvetica".}

\item{font_size_title}{Font size for the title text. Defaults to 11.}

\item{font_size_body}{Font size for all text other than the title. Defaults to 10.}
}
\value{
A ggplot theme.
}
\description{
Theme mobile adjustment for horizontal bar ggplots.
}
\examples{
library(ggplot2)
ggplot() +
theme_hbar("Courier", 9, 7) +
theme_hbar_mobile_adj("Courier", 9, 7)
}
