% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_map}
\alias{theme_map}
\title{Theme for maps.}
\usage{
theme_map(font_family = "", font_size_title = 11, font_size_body = 10)
}
\arguments{
\item{font_family}{Font family to use. Defaults to "".}

\item{font_size_title}{Font size for the title text. Defaults to 11.}

\item{font_size_body}{Font size for all text other than the title. Defaults to 10.}
}
\value{
A ggplot theme.
}
\description{
#' @title Theme for graphs with horizontal gridlines.
#' @param font_family Font family to use. Defaults to "".
#' @param font_size_title Font size for the title text. Defaults to 11.
#' @param font_size_body Font size for all text other than the title. Defaults to 10.
#' 
#' @return A ggplot theme.
#' @keywords internal
#'    
theme_h_gridlines <-
  function(font_family = "",
           font_size_title = 11,
           font_size_body = 10) {
    list(
      theme(
        text = element_text(family = font_family, size = font_size_body, colour = "#323232", face = "plain"),
        plot.title = element_text(size = font_size_title, face = "bold", colour = "#000000", hjust = 0.5, vjust = 1, margin = margin(b = font_size_body / 2)),
        plot.subtitle = element_text(colour = "#000000", hjust = 0.5, vjust = 1, margin = margin(b = font_size_body / 2)),
        axis.title.x = element_text(margin = margin(t = 10)),
        axis.title.y = element_text(angle = 90, margin = margin(r = 10)),
        legend.title = element_text(hjust = 0, margin = margin(r = 20)),
        plot.caption = element_text(hjust = 0.99, margin = margin(t = font_size_body)),
        axis.text.x = element_text(size = font_size_body, margin = margin(t = 2)),
        axis.text.y = element_text(hjust = 1, size = font_size_body, margin = margin(r = 2)),
        strip.text = element_text(hjust = 0.5, margin = margin(b = font_size_body / 2)),
        legend.text = element_text(margin = margin(r = 10), hjust = 0),
        plot.margin = margin(t = 5, l = 5, b = 5, r = 20),
        plot.caption.position = "plot",
        panel.border = element_blank(),
        panel.spacing = unit(2.5, "lines"),
        panel.grid.minor = element_blank(),
        panel.grid.major.x = element_blank(),
        panel.grid.major.y = element_line(colour = "#D3D3D3", size = 0.2),
        panel.background = element_rect(colour = "transparent", fill = "transparent"),
        strip.background = element_rect(colour = "transparent", fill = "transparent"),
        axis.line = element_line(colour = "#323232", size = 0.3),
        axis.ticks = element_line(colour = "#323232", size = 0.3),
        legend.margin = margin(t = 20, b = 20),
        legend.key = element_rect(colour = "transparent", fill = "transparent"),
        legend.key.height = unit(5, "mm"),
        legend.key.width = unit(5, "mm"),
        legend.direction = "vertical", 
        legend.box = NULL,
        complete = TRUE
      )
    )
  }
}
\details{
#' @title Theme for graphs with vertical gridlines.
#' @param font_family Font family to use. Defaults to "".
#' @param font_size_title Font size for the title text. Defaults to 11.
#' @param font_size_body Font size for all text other than the title. Defaults to 10.
#' 
#' @return A ggplot theme.
#' @keywords internal
#' 
theme_v_gridlines <-
  function(font_family = "",
           font_size_title = 11,
           font_size_body = 10) {
    list(
      theme(
        text = element_text(family = font_family, size = font_size_body, colour = "#323232", face = "plain"),
        plot.title = element_text(size = font_size_title, face = "bold", colour = "#000000", hjust = 0.5, vjust = 1, margin = margin(b = font_size_body / 2)),
        plot.subtitle = element_text(colour = "#000000", hjust = 0.5, vjust = 1, margin = margin(b = font_size_body / 2)),
        axis.title.x = element_text(margin = margin(t = 10)),
        axis.title.y = element_text(angle = 90, margin = margin(r = 10)),
        legend.title = element_text(hjust = 0, margin = margin(r = 20)),
        plot.caption = element_text(hjust = 0.99, margin = margin(t = font_size_body)),
        axis.text.x = element_text(size = font_size_body, margin = margin(t = 2)),
        axis.text.y = element_text(hjust = 1, size = font_size_body, margin = margin(r = 2)),
        strip.text = element_text(hjust = 0.5, margin = margin(b = font_size_body / 2)),
        legend.text = element_text(margin = margin(r = 10), hjust = 0),
        plot.margin = margin(t = 5, l = 5, b = 5, r = 20),
        plot.caption.position = "plot",
        panel.border = element_blank(),
        panel.spacing = unit(2.5, "lines"),
        panel.grid.minor = element_blank(),
        panel.grid.major.x = element_line(colour = "#D3D3D3", size = 0.2),
        panel.grid.major.y = element_blank(),
        panel.background = element_rect(colour = "transparent", fill = "transparent"),
        strip.background = element_rect(colour = "transparent", fill = "transparent"),
        axis.line = element_line(colour = "#323232", size = 0.3),
        axis.ticks = element_line(colour = "#323232", size = 0.3),
        legend.margin = margin(t = 20, b = 20),
        legend.key = element_rect(colour = "transparent", fill = "transparent"),
        legend.key.height = unit(5, "mm"),
        legend.key.width = unit(5, "mm"),
        legend.direction = "vertical",
        legend.box = NULL,
        complete = TRUE
      )
    )
  }

#' @title Theme for graphs with horizontal and vertical gridlines.
#' @param font_family Font family to use. Defaults to "".
#' @param font_size_title Font size for the title text. Defaults to 11.
#' @param font_size_body Font size for all text other than the title. Defaults to 10.
#' 
#' @return A ggplot theme.
#' @keywords internal
#' 
theme_hv_gridlines <-
  function(font_family = "",
           font_size_title = 11,
           font_size_body = 10) {
    list(
      theme(
        text = element_text(family = font_family, size = font_size_body, colour = "#323232", face = "plain"),
        plot.title = element_text(size = font_size_title, face = "bold", colour = "#000000", hjust = 0.5, vjust = 1, margin = margin(b = font_size_body / 2)),
        plot.subtitle = element_text(colour = "#000000", hjust = 0.5, vjust = 1, margin = margin(b = font_size_body / 2)),
        axis.title.x = element_text(margin = margin(t = 10)),
        axis.title.y = element_text(angle = 90, margin = margin(r = 10)),
        legend.title = element_text(hjust = 0, margin = margin(r = 20)),
        plot.caption = element_text(hjust = 0.99, margin = margin(t = font_size_body)),
        axis.text.x = element_text(size = font_size_body, margin = margin(t = 2)),
        axis.text.y = element_text(hjust = 1, size = font_size_body, margin = margin(r = 2)),
        strip.text = element_text(hjust = 0.5, margin = margin(b = font_size_body / 2)),
        legend.text = element_text(margin = margin(r = 10), hjust = 0),
        plot.margin = margin(t = 5, l = 5, b = 5, r = 20),
        plot.caption.position = "plot",
        panel.border = element_blank(),
        panel.spacing = unit(2.5, "lines"),
        panel.grid.minor = element_blank(),
        panel.grid.major.x = element_line(colour = "#D3D3D3", size = 0.2),
        panel.grid.major.y = element_line(colour = "#D3D3D3", size = 0.2),
        panel.background = element_rect(colour = "transparent", fill = "transparent"),
        strip.background = element_rect(colour = "transparent", fill = "transparent"),
        axis.line = element_line(colour = "#323232", size = 0.3),
        axis.ticks = element_line(colour = "#323232", size = 0.3),
        legend.margin = margin(t = 20, b = 20),
        legend.key = element_rect(colour = "transparent", fill = "transparent"),
        legend.key.height = unit(5, "mm"),
        legend.key.width = unit(5, "mm"),
        legend.direction = "vertical",
        legend.box = NULL,
        complete = TRUE
      )
    )
  }

#' @title Theme for graphs with no gridlines.
#' @param font_family Font family to use. Defaults to "".
#' @param font_size_title Font size for the title text. Defaults to 11.
#' @param font_size_body Font size for all text other than the title. Defaults to 10.
#' 
#' @return A ggplot theme.
#' @keywords internal
#' 
theme_no_gridlines <-
  function(font_family = "",
           font_size_title = 11,
           font_size_body = 10) {
    list(
      theme(
        text = element_text(family = font_family, size = font_size_body, colour = "#323232", face = "plain"),
        plot.title = element_text(size = font_size_title, face = "bold", colour = "#000000", hjust = 0.5, vjust = 1, margin = margin(b = font_size_body / 2)),
        plot.subtitle = element_text(colour = "#000000", hjust = 0.5, vjust = 1, margin = margin(b = font_size_body / 2)),
        axis.title.x = element_text(margin = margin(t = 10)),
        axis.title.y = element_text(angle = 90, margin = margin(r = 10)),
        legend.title = element_text(hjust = 0, margin = margin(r = 20)),
        plot.caption = element_text(hjust = 0.99, margin = margin(t = font_size_body)),
        axis.text.x = element_text(size = font_size_body, margin = margin(t = 2)),
        axis.text.y = element_text(hjust = 1, size = font_size_body, margin = margin(r = 2)),
        strip.text = element_text(hjust = 0.5, margin = margin(b = font_size_body / 2)),
        legend.text = element_text(margin = margin(r = 10), hjust = 0),
        plot.margin = margin(t = 5, l = 5, b = 5, r = 20),
        plot.caption.position = "plot",
        panel.border = element_blank(),
        panel.spacing = unit(2.5, "lines"),
        panel.grid.minor = element_blank(),
        panel.grid.major.x = element_blank(),
        panel.grid.major.y = element_blank(),
        panel.background = element_rect(colour = "transparent", fill = "transparent"),
        strip.background = element_rect(colour = "transparent", fill = "transparent"),
        axis.line = element_line(colour = "#323232", size = 0.3),
        axis.ticks = element_line(colour = "#323232", size = 0.3),
        legend.margin = margin(t = 20, b = 20),
        legend.key = element_rect(colour = "transparent", fill = "transparent"),
        legend.key.height = unit(5, "mm"),
        legend.key.width = unit(5, "mm"),
        legend.direction = "vertical", 
        legend.box = NULL,
        complete = TRUE
      )
    )
  }
}
\keyword{internal}
