% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_theme.R
\name{theme_mobile_extra}
\alias{theme_mobile_extra}
\title{Extra theme elements for ggplot2 graphs on mobile devices}
\usage{
theme_mobile_extra(void = FALSE)
}
\arguments{
\item{void}{TRUE or FALSE of whether to drop all axis lines, ticks and x and y labels. Useful for maps. Defaults to FALSE.}
}
\value{
A ggplot theme.
}
\description{
Extra theme elements for ggplot2 graphs on mobile devices
}
\keyword{internal}
