% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_mobile_extra_map}
\alias{theme_mobile_extra_map}
\title{Extra theme elements for ggplot2 maps on mobile devices}
\usage{
theme_mobile_extra_map()
}
\value{
A ggplot theme.
}
\description{
Extra theme elements for ggplot2 maps on mobile devices
}
\keyword{internal}
