% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_mobile_graph}
\alias{theme_mobile_graph}
\title{Extra theme elements for ggplot2 graphs on mobile devices}
\usage{
theme_mobile_graph()
}
\value{
A ggplot theme.
}
\description{
Extra theme elements for ggplot2 graphs on mobile devices
}
\keyword{internal}
