% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breaks.R
\name{y_numeric_breaks}
\alias{y_numeric_breaks}
\title{Calculate the breaks for a y axis numeric vector.}
\usage{
y_numeric_breaks(y_var_vctr, y_balance = FALSE, y_pretty_n = 5,
  y_trans = "identity", y_zero = TRUE)
}
\arguments{
\item{y_var_vctr}{A numeric vector for the y scale from which to determine breaks from.}

\item{y_balance}{Add balance to the y axis so that zero is in the centre of the y scale.}

\item{y_pretty_n}{The desired number of intervals on the y axis, as calculated by the pretty algorithm. Defaults to 5.}

\item{y_trans}{A string specifying a transformation for the y axis scale, such as "log10" or "sqrt". Defaults to "identity".}

\item{y_zero}{TRUE or FALSE of whether the minimum of the y scale is zero. Defaults to TRUE.}
}
\value{
A vector of breaks
}
\description{
Calculate the breaks for a y axis numeric vector.
}
