% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_sf_polygon}
\alias{example_sf_polygon}
\title{Example sf polygon object.}
\format{
An \code{sf} object.
}
\usage{
example_sf_polygon
}
\description{
Example sf polygon object.
}
\examples{
example_sf_polygon
}
\keyword{datasets}
