% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_stars}
\alias{example_stars}
\title{Example stars object.}
\format{An \code{stars} object.}
\usage{
example_stars
}
\description{
Example stars object.
}
\examples{
example_stars
}
\keyword{datasets}
