% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{gg_theme}
\alias{gg_theme}
\title{Create a simple theme for a graph.}
\usage{
gg_theme(
  family = "",
  size_title = 11,
  size_subtitle = 10,
  size_body = 10,
  size_axis = 0.3,
  size_ticks = 0.3,
  size_gridlines = 0.2,
  face_title = "bold",
  face_subtitle = "plain",
  face_body = "plain",
  pal_title = "#000000",
  pal_subtitle = "#000000",
  pal_body = "#323232",
  pal_axis = "#323232",
  pal_ticks = "#323232",
  pal_gridlines = "#D3D3D3",
  gridlines = "horizontal"
)
}
\arguments{
\item{family}{The font for all text to use. Defaults to "".}

\item{size_title}{The size of the title font. Defaults to 11.}

\item{size_subtitle}{The size of the subtitle font. Defaults to 10.}

\item{size_body}{The size of all font other than the title or subtitle. Defaults to 10.}

\item{size_axis}{The size of the axis. Defaults to 0.3.}

\item{size_ticks}{The size of the ticks. Defaults to 0.3.}

\item{size_gridlines}{The size of the vertical major gridlines. Defaults to 0.2.}

\item{face_title}{The face of the title font. Defaults to "bold".}

\item{face_subtitle}{The face of the subtitle font. Defaults to "plain".}

\item{face_body}{The face of all font other than the title or subtitle. Defaults to "plain".}

\item{pal_title}{The colour palette for the title font. Defaults to "#000000".}

\item{pal_subtitle}{The colour palette for the subtitle font. Defaults to "#000000".}

\item{pal_body}{The colour palette for all font other than the title or subtitle. Defaults to "#323232".}

\item{pal_axis}{The colour palette for the axis. Defaults to "#323232".}

\item{pal_ticks}{The colour palette for the ticks. Defaults to "#323232".}

\item{pal_gridlines}{The colour palette for the vertical major gridlines. Defaults to "#D3D3D3".}

\item{gridlines}{Whether gridlines are "horizontal", "vertical", "both" or "none".}
}
\value{
A ggplot theme.
}
\description{
Create a simple theme for a graph.
}
