% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{gg_theme_map}
\alias{gg_theme_map}
\title{Create a simple theme for a map.}
\usage{
gg_theme_map(
  family = "",
  size_title = 11,
  size_subtitle = 10,
  size_body = 10,
  face_title = "bold",
  face_subtitle = "plain",
  face_body = "plain",
  pal_title = "#000000",
  pal_subtitle = "#000000",
  pal_body = "#323232"
)
}
\arguments{
\item{family}{The font for all text to use. Defaults to "".}

\item{size_title}{The size of the title font. Defaults to 11.}

\item{size_subtitle}{The size of the subtitle font. Defaults to 10.}

\item{size_body}{The size of all font other than the title or subtitle. Defaults to 10.}

\item{face_title}{The face of the title font. Defaults to "bold".}

\item{face_subtitle}{The face of the subtitle font. Defaults to "plain".}

\item{face_body}{The face of all font other than the title or subtitle. Defaults to "plain".}

\item{pal_title}{The colour palette for the title font. Defaults to "#000000".}

\item{pal_subtitle}{The colour palette for the subtitle font. Defaults to "#000000".}

\item{pal_body}{The colour palette for all font other than the title or subtitle. Defaults to "#323232".}
}
\value{
A ggplot theme.
}
\description{
Create a simple theme for a map.
}
\keyword{internal}
