% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaf_clear.R
\name{leaf_clear}
\alias{leaf_clear}
\title{In shiny, clear all features, images and legends.}
\usage{
leaf_clear(map_id = "leaf")
}
\arguments{
\item{map_id}{The map id for a leaflet map. Defaults to "leaf".}
}
\value{
A map object.
}
\description{
In shiny, clear all features, images and legends.
}
