% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_basemap_fun.R
\name{leaflet_basemap}
\alias{leaflet_basemap}
\title{Basemap stack in leaflet.}
\usage{
leaflet_basemap(top_layer = "light", country = NULL, state = NULL,
  continent = NULL)
}
\arguments{
\item{top_layer}{The first layer to start in the basemap stack. Either "light", "dark", "street", "satellite", or "ocean". Defaults to "light".}

\item{country}{A country .data$name from Natural Earth.}

\item{state}{A state or region .data$name within a country from Natural Earth. Must be used in combination with the applicable country specified.}

\item{continent}{A continent .data$name from Natural Earth.}
}
\value{
A leaflet object.
}
\description{
Make a stack of leaflet baselayers for use in shiny apps.
}
\examples{
leaflet_basemap("dark")
leaflet_basemap(country = "Papua New Guinea")
}
