% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_text.R
\name{mutate_text}
\alias{mutate_text}
\title{Add a quick tooltip text column to data.}
\usage{
mutate_text(data, text_vars_vctr = NULL, comma = FALSE)
}
\arguments{
\item{data}{A tibble or dataframe. Required input.}

\item{text_vars_vctr}{A vector of quoted variables to include in the tooltip. Defaults to NULL, which adds all variables in.}

\item{comma}{TRUE or FALSE of whether to convert numeric values to character values with comma seperators.}
}
\value{
A vector of labels.
}
\description{
Add a column of tooltip text which is automatically created based on column names and values.
}
\examples{
library(dplyr)

plot_data <- slice_sample(ggplot2::diamonds, prop = 0.05) \%>\% 
  mutate_text(c("carat", "price"), comma = TRUE)

plot <- gg_point(data = plot_data, 
                 x_var = carat, 
                 y_var = price, 
                 text_var = text, 
                 title = "Diamond price by carat", 
                 x_title = "Carat", 
                 y_title = "Price ($US thousands)")

plotly::ggplotly(plot, tooltip = "text")

}
