% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nz}
\alias{nz}
\title{New Zealand coastline.}
\format{An \code{sf} object.}
\usage{
nz
}
\description{
Simplified New Zealand coastline boundary, excluding the Chatham Islands.
}
\examples{
nz

ggplot_sf(nz)

ggplot_sf(dplyr::slice(nz, 2, 4))

ggplot_sf(dplyr::slice(nz, 1, 3, 5:7))
}
\keyword{datasets}
