% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nz_region}
\alias{nz_region}
\title{New Zealand clipped regional boundaries}
\format{An \code{sf} object.}
\usage{
nz_region
}
\description{
Simplified New Zealand regional boundaries, excluding the Chatham Islands.
}
\examples{
nz_region

ggplot_sf(nz_region)

ggplot_sf(dplyr::filter(nz_region, region == "Wellington"))
}
\keyword{datasets}
