% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.R
\name{pal_na}
\alias{pal_na}
\title{NA palette.}
\usage{
pal_na(col_n = 1)
}
\arguments{
\item{col_n}{The number of NA colours. Doesn't do anything, but provided for consistency.}
}
\value{
A vector of one hex code.
}
\description{
A function to retreive a hex code for a colour to use for NA values.
}
\examples{
scales::show_col(pal_na(9)) 
}
