% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/er_helpers.R, R/other.R
\docType{data}
\name{pal_point_set1}
\alias{pal_point_set1}
\title{Colour palette for categorical variables for points.}
\format{An object of class \code{character} of length 9.}
\usage{
pal_point_set1

pal_point_set1
}
\value{
A vector of hex codes.

A vector of hex codes.
}
\description{
Colour palette for categorical variables.

Colour palette for categorical variables.
}
\keyword{datasets}
