% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_template}
\alias{run_template}
\title{Run shiny template with option to download.}
\usage{
run_template(template = "template1", ...)
}
\arguments{
\item{template}{template name. Available templates are "template1" with graph and table tabs, and "template2" also with a map tab. Defaults to "template1".}

\item{...}{passed to \code{shiny::runApp}}
}
\description{
Run shiny template with option to download.
}
