% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_boxplot_stats.R
\name{summarise_boxplot_stats}
\alias{summarise_boxplot_stats}
\title{Summarise boxplot stats}
\usage{
summarise_boxplot_stats(data, var, ...)
}
\arguments{
\item{data}{A tibble or dataframe. Required input.}

\item{var}{A numeric variable. Required input.}

\item{...}{Passed to \code{boxplot.stats}}
}
\value{
A tibble or dataframe.
}
\description{
Summarise boxplot stats
}
\examples{
library(dplyr)
library(simplevis)
library(palmerpenguins)

penguins \%>\% 
  group_by(sex, species) \%>\% 
  summarise_boxplot_stats(body_mass_g) 

penguins \%>\% 
  group_by(sex, species) \%>\% 
  summarise_boxplot_stats(body_mass_g) \%>\% 
  gg_boxplot_col(sex, body_mass_g, species, stat = "identity")
  
}
