% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_max.R
\name{sv_density_max}
\alias{sv_density_max}
\title{Calculate the max density in a gg_density() plot.}
\usage{
sv_density_max(
  data,
  x_var,
  stat_bw = "nrd0",
  stat_adjust = 1,
  stat_kernel = "gaussian",
  stat_n = 512,
  stat_trim = FALSE
)
}
\arguments{
\item{data}{A tibble or dataframe. Required input.}

\item{x_var}{Unquoted numeric variable to be on the x scale. Required input.}

\item{stat_bw}{The bw argument of the stats::density function. Defaults to "nrd0".}

\item{stat_adjust}{The adjust argument of the stats::density function. Defaults to 1.}

\item{stat_kernel}{The kernel argument of the stats::density function. Defaults to "gaussian".}

\item{stat_n}{The n argument of the stats::density function. Defaults to 512.}

\item{stat_trim}{The trim argument of the stats::density function. Defaults to FALSE.}
}
\description{
Calculate the max density in a gg_density() plot.
}
\keyword{internal}
