% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_max.R
\name{sv_density_max_facet}
\alias{sv_density_max_facet}
\title{Calculate the max density in a gg_density_facet() plot.}
\usage{
sv_density_max_facet(
  data,
  x_var,
  facet_var,
  algo_bw = "nrd0",
  algo_adjust = 1,
  algo_kernel = "gaussian",
  algo_n = 512,
  algo_trim = FALSE
)
}
\arguments{
\item{data}{A tibble or dataframe. Required input.}

\item{x_var}{Unquoted numeric variable to be on the x scale. Required input.}

\item{facet_var}{Unquoted categorical variable to facet by. Required input.}

\item{algo_bw}{The bw argument of the stats::density function. Defaults to "nrd0".}

\item{algo_adjust}{The adjust argument of the stats::density function. Defaults to 1.}

\item{algo_kernel}{The kernel argument of the stats::density function. Defaults to "gaussian".}

\item{algo_n}{The n argument of the stats::density function. Defaults to 512.}

\item{algo_trim}{The trim argument of the stats::density function. Defaults to FALSE.}
}
\description{
Calculate the max density in a gg_density_facet() plot.
}
\keyword{internal}
