% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_labels.R
\name{sv_interval_labels_num}
\alias{sv_interval_labels_num}
\title{Convert numeric bin cuts to simple and pretty labels.}
\usage{
sv_interval_labels_num(cuts, digits = NULL, right_closed = TRUE)
}
\arguments{
\item{cuts}{A vector of numeric cuts.}

\item{digits}{If cuts are numeric, the number of decimal places to round labels to.}

\item{right_closed}{If cuts are numeric, TRUE or FALSE of whether intervals are to be right-closed. Defaults to TRUE.}
}
\value{
A vector of character labels.
}
\description{
Convert numeric bin cuts to simple and pretty labels.
}
\examples{
library(simplevis)

interval_labels(c(0, 0.1, 3, 4.1, 7, 100, Inf))
}
\keyword{internal}
